refactor(apps-proxy): AJDA-3219 extract the framework background-poll check
Streamlit emits /_stcore/health and /_stcore/host-config on every
websocket reconnect, independently of the user, and the upstream handler
already excludes them from the auto-suspend notification. Session
tracking needs the same rule, so "the user did something" has to have one
definition — and it cannot live in the upstream package, which will
depend on the sessions package in the other direction.

Behaviour is unchanged; the function only moved.
