+++ internal/pkg/service/appsproxy/dataapps/frameworkpoll/frameworkpoll_test.go
@@ -1,4 +1,4 @@
-package upstream
+package frameworkpoll
 
 import (
 	"testing"
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestIsFrameworkBackgroundPoll(t *testing.T) {
+func TestIs(t *testing.T) {
 	t.Parallel()
 	tests := []struct {
 		path string
@@ -35,7 +35,7 @@ func TestIsFrameworkBackgroundPoll(t *testing.T) {
 	for _, tc := range tests {
 		t.Run(tc.path, func(t *testing.T) {
 			t.Parallel()
-			assert.Equal(t, tc.want, isFrameworkBackgroundPoll(tc.path))
+			assert.Equal(t, tc.want, Is(tc.path))
 		})
 	}
 }
